feat: Classify a given flower
Once you can get images in the correct format, it's time to write a function for making predictions with your model. A common practice is to predict the top 5 or so (usually called top- 𝐾 ) most probable classes. You'll want to calculate the class probabilities then find the  𝐾  largest values.

To get the top K largest values in a tensor use x.topk(k). This method returns both the highest k probabilities and the indices of those probabilities corresponding to the classes. You need to convert from these indices to the actual class labels using class_to_idx which hopefully you added to the model or from an ImageFolder you used to load the data. Make sure to invert the dictionary so you get a mapping from index to class as well.