feat: Do sanity checking
After the CNN has been trained it is good practice to make sure that
there aren't obvious bugs. In this part we will print the flower and plot
a barplot with the most likely 5 classes it belongs to. We will also add
the real name of the flower (from the label) to the title of the plot.